LayoutTests/http/tests/security/document-all-expected.txt

CONSOLE MESSAGE: line 1: TypeError: WindowConstructor is not a function (evaluating 'obj.alert.constructor("return document.body.innerHTML")')
